+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.0.2] - 2024-05-01
+
+### Changed 
+- `dsf-email-templater` use `fs/promises` to load and save files
+- `mailtrap-njk` use `fs/promises` to load files
+
+### Removed
+- Removed use of `fs/promises`, `saveFile()` and `renderFromFile()` from `DSFEmailRenderer`
+
 ## [v2.0.1] - 2024-04-30
 
 ### Fixed